# packages/openapi/src/openapi.yaml
openapi: 3.0.3
info:
  title: PeridotID API
  version: 0.1.0
  description: |
    Gaming Identity Platform — Authentication, Identity, and Profile for the Peridot ecosystem.

    ## Base URL
    All endpoints are prefixed with `/v1`.

    | Environment | Base URL |
    |---|---|
    | Production | `https://api.pid.peridotvault.com/v1` |
    | Local dev | `http://localhost:3301/v1` |

    ## Authentication
    Authentication is **cookie-based**. After a successful Google sign-in the API sets two
    `HttpOnly` cookies:

    - `pid_access` — short-lived JWT access token (15 min), sent on every request.
    - `pid_refresh` — long-lived rotating refresh token (30 days), only sent to `/v1/auth`.

    Browsers handle these automatically; non-browser clients must send
    `Cookie: pid_access=<token>` on each request and call `/v1/auth/refresh` before the
    access token expires.
  contact:
    name: PeridotID
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.pid.peridotvault.com/v1
    description: Production
  - url: http://localhost:3301/v1
    description: Local development
...
paths:
...
  /fiat/deposits:
    get:
      tags: [fiat]
      summary: Recent deposit intents for the caller (money-in history)
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Deposit intents
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
        "401":
          $ref: "#/components/responses/Unauthorized"
  /fiat/deposits/checkout:
    post:
      tags: [fiat]
      summary: Create a Checkout deposit intent (DOKU-hosted page, all channels)
      description: |
        netAmountIdr is the NET credited to the user (minimum Rp100.000,
        API-enforced — no DOKU minimum). Global PeridotID fee = 0.1% of net,
        min Rp100, no cap; when `clientId` is given the app's topup fee stacks
        on top. gross = net + fee, quoted upfront and snapshotted on the row.
        DOKU charges its own provider fee separately.
        The page is routed to the caller's sub-account. The customer is
        charged gross; fiat settles 100% to the user IDR account as points
        backing. On verified payment the backend issues NET points (user)
        + FEE points (Treasury) on the Unified Ledger — the Saldo is usable
        in seconds without waiting for fiat settlement. The API never debits
        a fee after settlement.
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [netAmountIdr]
              properties:
                netAmountIdr:
                  type: string
      responses:
        "200":
          description: Checkout deposit intent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutDeposit"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /fiat/fee-policy:
    get:
      tags: [fiat]
      summary: Active platform service-fee policy
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Fee policy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeePolicy"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /fiat/deposits/{id}/sync:
    post:
      tags: [fiat]
      summary: Reconcile a created/processing row via DOKU transactions-status
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubTx"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /fiat/webhook:
    post:
      tags: [fiat]
      summary: Sub-Account notification inbox (persist-first, corroborated before apply)
      description: |
        No published V2 webhook signing contract — events persist on X-EXTERNAL-ID
        and apply only after DOKU transactions-status corroboration. Public; always 200 after persist.
      parameters:
        - name: X-EXTERNAL-ID
          in: header
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Acknowledged
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    example: true
  /fiat/admin/webhooks:
    get:
      tags: [fiat]
      summary: Webhook inbox read-model (admin)
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Inbox events
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /fiat/admin/backfill:
    post:
      tags: [fiat]
      summary: Backfill makeup points for settled deposits missing issuance (admin)
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                pid:
                  type: string
                take:
                  type: integer
      responses:
        "200":
          description: Backfill report
          content:
            application/json:
              schema:
                type: object
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /fiat/admin/fee-policy:
    post:
      tags: [fiat]
      summary: Publish a new fee-policy version (admin, deactivates previous)
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [percentBps, minIdr, maxIdr]
              properties:
                percentBps:
                  type: integer
                minIdr:
                  type: string
                maxIdr:
                  type: string
      responses:
        "200":
          description: Fee policy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeePolicy"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /fiat/balance:
    get:
      tags: [fiat]
      summary: Internal-credit balance (temporary Rail B, replayed — source labelled)
      description: |
        Temporary Internal Credit Ledger balance, replayed from the immutable
        journal (source: internal-credit — NOT the DOKU POINT Saldo).
        See docs/INTERNAL_CREDIT.md.
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Internal credit balance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FiatBalance"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/Unavailable"
  /fiat/ledger:
    get:
      tags: [fiat]
      summary: Internal-credit journal export + replay verdict
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Journal rows and replay verdict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FiatLedger"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /fiat/transfers/inquiry:
    post:
      tags: [fiat]
      summary: Internal-credit send inquiry (moves no money, phase-1 policy enforced)
      description: |
        GROSS-in quote pinned by feePolicyVersion. Phase-1 policy (escrow-only):
        one side of every send must be an allowlisted app PID.
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amountIdr, beneficiaryPid]
              properties:
                amountIdr:
                  type: string
                beneficiaryPid:
                  type: string
                remark:
                  type: string
      responses:
        "200":
          description: Transfer inquiry with server quote
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FiatTransferInquiry"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /fiat/transfers/{id}/confirm:
    post:
      tags: [fiat]
      summary: Confirm an internal-credit send (atomic, idempotent)
      description: |
        Posts sender-out + recipient-in + treasury-fee legs in one DB
        transaction. Re-confirming a posted intent returns the view.
      security:
        - cookieAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Posted transfer view
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FiatLedgerEntry"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /fiat/transfers/{id}/cancel:
    post:
      tags: [fiat]
      summary: Cancel a created internal-credit intent
      security:
        - cookieAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Transfer view
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FiatLedgerEntry"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /fiat/admin/replay:
    get:
      tags: [fiat]
      summary: Platform-wide internal-credit replay projection (admin)
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Replay projection and rail switches
          content:
            application/json:
              schema:
                type: object
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /fiat/admin/chain:
    get:
      tags: [fiat]
      summary: Verify the tamper-evident ledger hash chain (admin)
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Chain verification verdict + head hash
          content:
            application/json:
              schema:
                type: object
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /fiat/admin/chain/head:
    get:
      tags: [fiat]
      summary: Current chain head hash (publish externally to anchor) (admin)
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Head hash + seq
          content:
            application/json:
              schema:
                type: object
        "401":
          $ref: "#/components/responses/Unauthorized"
  /fiat/admin/chain/backfill:
    post:
      tags: [fiat]
      summary: Chain posted rows that predate the chain (admin)
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Backfill result
          content:
            application/json:
              schema:
                type: object
        "401":
          $ref: "#/components/responses/Unauthorized"
  /fiat/admin/dispatch-events:
    post:
      tags: [fiat]
      summary: Deliver due third-party callbacks (cron) (admin)
      security:
        - cookieAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                take:
                  type: integer
      responses:
        "200":
          description: Dispatch result
          content:
            application/json:
              schema:
                type: object
        "401":
          $ref: "#/components/responses/Unauthorized"
  /fiat/admin/events:
    get:
      tags: [fiat]
      summary: Inspect the callback outbox (admin)
      security:
        - cookieAuth: []
      parameters:
        - in: query
          name: status
          schema:
            type: string
      responses:
        "200":
          description: Outbox events
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
        "401":
          $ref: "#/components/responses/Unauthorized"
  /fiat/admin/freeze:
    post:
      tags: [fiat]
      summary: Engage/clear the internal-credit migration freeze (admin)
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [frozen]
              properties:
                frozen:
                  type: boolean
                reason:
                  type: string
      responses:
        "200":
          description: Freeze state
          content:
            application/json:
              schema:
                type: object
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: pid_access
      description: "JWT access token set by /auth/google/callback or /auth/refresh"
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
          example:
            statusCode: 400
            message: ["handle must be 3-20 chars: lowercase letters, numbers, underscore"]
...
    Unauthorized:
      description: Missing, expired, or invalid access/refresh token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
          example:
            statusCode: 401
            message: "Unauthorized"
    Forbidden:
      description: Authenticated but lacking the required role (e.g. admin)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
          example:
            statusCode: 403
            message: "Admin access required"
    NotFound:
      description: The identity or profile does not exist
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
          example:
            statusCode: 404
            message: "No Identity found"
...
  schemas:
...
    SubTx:
      type: object
      required: [id, kind, providerRef, providerStatus, grossIdr, currency, createdAt]
      properties:
        id:
          type: string
        kind:
          type: string
          description: deposit | payout_bank | payout_wallet | transfer_internal | debit | debit_cancel | fee | points_issue | points_fee | points_clawback
        providerRef:
          type: string
          description: Idempotency key (partnerReferenceNo at DOKU)
        providerStatus:
          type: string
          enum: [created, processing, settled, success, expired, failed, cancelled, refunded]
          description: Payment state (never settlement). Checkout deposits use success/expired/failed; legacy rows may carry settled meaning paid.
        grossIdr:
          type: string
          description: Gross movement (Checkout deposit = customer-paid net + fee; transfer = requested gross, beneficiary gets gross minus fee)
        feeIdr:
          type: string
          nullable: true
          description: Platform service fee quote, snapshotted at intent time (legacy fee rows remain for history)
        netIdr:
          type: string
          nullable: true
          description: Net user effect (gross minus fee)
        paymentUrl:
          type: string
          nullable: true
          description: DOKU-hosted payment page for unpaid Checkout deposits
        feeStatus:
          type: string
          nullable: true
          description: Legacy fee-leg status ({ref}-FEE row), read-only
        currency:
          type: string
          enum: [IDR]
        createdAt:
          type: string
          format: date-time
    FiatBalance:
      type: object
      required: [balanceIdr, currency, source]
      properties:
        balanceIdr:
          type: string
          description: Replayed fiat-ledger balance, whole IDR
        currency:
          type: string
          enum: [IDR]
        source:
          type: string
          enum: [fiat-ledger]
        replayErrors:
          type: array
          items:
            type: string
    FiatLedgerEntry:
      type: object
      required: [id, kind, entryGroup, grossIdr, status, createdAt]
      properties:
        id:
          type: string
        kind:
          type: string
          description: fiat_issue | fiat_fee | fiat_transfer_in | fiat_transfer_out | fiat_adjust
        entryGroup:
          type: string
          description: Groups the legs of one logical movement
        counterpartyPid:
          type: string
          nullable: true
        grossIdr:
          type: string
        feeIdr:
          type: string
          nullable: true
        netIdr:
          type: string
          nullable: true
        status:
          type: string
          enum: [created, posted, failed, cancelled]
        createdAt:
          type: string
          format: date-time
    FiatTransferInquiry:
      type: object
      required: [id, entryGroup, grossIdr, feeIdr, netIdr, feePolicyVersion, beneficiaryPid]
      properties:
        id:
          type: string
        entryGroup:
          type: string
        grossIdr:
          type: string
        feeIdr:
          type: string
        netIdr:
          type: string
        feePolicyVersion:
          type: integer
        beneficiaryPid:
          type: string
    FiatLedger:
      type: object
      properties:
        pid:
          type: string
        balanceIdr:
          type: string
        treasuryCreditedIdr:
          type: string
        inFlight:
          type: array
          items:
            type: string
        errors:
          type: array
          items:
            type: string
        rows:
          type: array
          items:
            type: object
...
    CheckoutDeposit:
      type: object
      required: [id, providerRef, paymentUrl, tokenId, grossIdr, feeIdr, netIdr, feePolicyVersion, providerStatus, createdAt]
      properties:
        id:
          type: string
        providerRef:
          type: string
          description: Invoice number at DOKU (idempotency key)
        paymentUrl:
          type: string
          description: DOKU-hosted payment page (all banks, QRIS, e-money, cards)
        tokenId:
          type: string
        expiredDate:
          type: string
          nullable: true
        grossIdr:
          type: string
          description: Gross charged at DOKU (net + fee)
        feeIdr:
          type: string
          description: Fee quoted on the entered net (honored at settlement)
        netIdr:
          type: string
          description: Net credited to the user (the entered amount)
        feePolicyVersion:
          type: integer
        providerStatus:
          type: string
        createdAt:
          type: string
          format: date-time
    FeePolicy:
      type: object
      required: [version, percentBps, minIdr, maxIdr, active]
      properties:
        version:
          type: integer
        percentBps:
          type: integer
          description: Basis points, e.g. 10 = 0.1%
        minIdr:
          type: string
          description: Fee floor in whole IDR (0 = no floor). Default Rp100
        maxIdr:
          type: string
          description: Fee cap in whole IDR (0 = uncapped). Default 0 (no cap)
        active:
          type: boolean
...
    ApiError:
      type: object
      required: [statusCode, message]
      properties:
        statusCode:
          type: integer
        message:
          oneOf:
            - type: string
            - type: array
              items:
                type: string